requirements.txt
scikit-learn
albumentations
tqdm
pandas
numpy
loguru
h5py
Pillow
six
scipy
opencv-python
matplotlib
yacs
gdown
imageio
tabulate
pycocotools
thop
easydict
natsort
Cython
open-clip-torch
transformers
timm
pyyaml
lapx
ptflops
einops
unsloth
unsloth_zoo
tensorboard
mplsoccer
ftfy
regex
cython_bbox
ninja
rich
# torchreid is often installed via: pip install git+https://github.com/KaiyangZhou/deep-person-reid.git
clip @ git+https://github.com/openai/CLIP.git